+++ app/src/main/java/com/alexeybondarenko/picsearch/ui/settings/SettingsScreen.kt
@@ -3,7 +3,7 @@ package com.alexeybondarenko.picsearch.ui.settings
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import com.alexeybondarenko.picsearch.ui.utils.ScreenPlaceholder
+import com.alexeybondarenko.picsearch.ui.utils.common.ScreenPlaceholder
 
 @Preview
 @Composable
